下一步准备把休眠加进去, 测试

+++ User/Inc/rfid.h
@@ -1,6 +1,6 @@
 #ifndef _RFID_H_
 #define _RFID_H_
 
-u8 reader_mode(void);
+int reader_mode(void);
 
 #endif // _RFID_H_
